🗑 Disable auto-lightbox on formats other than `amp/⚡` (#23228)
amp-lightbox-gallery is not allowed on amp4email, which makes the loading of this extension unnecessary and could lead to bugs.


This enables it only for `<html amp>` (or the cute version  `<html ⚡>`) and disables it everywhere else.

Proxy-origin check also moves from extension into core.
